move current samples to a folder named deprecated
because the functions will all be deprecated. We will keep those files
because:
- still valuable in case someone does not switch immediately to the new
  expectation functions (i.e. uses not the `to + infinitive` version).
  In such a case one sees the samples in the KDoc
- we can copy them once we have introduced the corresponding functions
