+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/deprecated/ComparableAssertionSamples.kt
@@ -1,10 +1,14 @@
-package ch.tutteli.atrium.api.fluent.en_GB.samples
+//TODO remove file with 1.0.0
+@file:Suppress("DEPRECATION")
+
+package ch.tutteli.atrium.api.fluent.en_GB.samples.deprecated
 
 import ch.tutteli.atrium.api.fluent.en_GB.isEqualComparingTo
 import ch.tutteli.atrium.api.fluent.en_GB.isGreaterThan
 import ch.tutteli.atrium.api.fluent.en_GB.isGreaterThanOrEqual
 import ch.tutteli.atrium.api.fluent.en_GB.isLessThan
 import ch.tutteli.atrium.api.fluent.en_GB.isLessThanOrEqual
+import ch.tutteli.atrium.api.fluent.en_GB.samples.fails
 import ch.tutteli.atrium.api.verbs.internal.expect
 import kotlin.test.Test
