+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/iteratorAssertions.kt
@@ -12,7 +12,7 @@ import ch.tutteli.atrium.logic.hasNotNext
  *
  * @since 0.13.0
  *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.IteratorAssertionSamples.has
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.deprecated.IteratorAssertionSamples.has
  */
 infix fun <E, T : Iterator<E>> Expect<T>.has(@Suppress("UNUSED_PARAMETER") next: next): Expect<T> =
     _logicAppend { hasNext() }
@@ -24,7 +24,7 @@ infix fun <E, T : Iterator<E>> Expect<T>.has(@Suppress("UNUSED_PARAMETER") next:
  *
  * @since 0.13.0
  *
- * @sample ch.tutteli.atrium.api.infix.en_GB.samples.IteratorAssertionSamples.hasNot
+ * @sample ch.tutteli.atrium.api.infix.en_GB.samples.deprecated.IteratorAssertionSamples.hasNot
  */
 infix fun <E, T : Iterator<E>> Expect<T>.hasNot(@Suppress("UNUSED_PARAMETER") next: next): Expect<T> =
     _logicAppend { hasNotNext() }
