+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/infix/en_GB/samples/deprecated/IteratorAssertionSamples.kt
@@ -1,8 +1,12 @@
-package ch.tutteli.atrium.api.infix.en_GB.samples
+//TODO remove file with 1.0.0
+@file:Suppress("DEPRECATION")
+
+package ch.tutteli.atrium.api.infix.en_GB.samples.deprecated
 
 import ch.tutteli.atrium.api.infix.en_GB.has
 import ch.tutteli.atrium.api.infix.en_GB.hasNot
 import ch.tutteli.atrium.api.infix.en_GB.next
+import ch.tutteli.atrium.api.infix.en_GB.samples.fails
 import ch.tutteli.atrium.api.verbs.internal.expect
 import kotlin.test.Test
