+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/infix/en_GB/samples/deprecated/ListAssertionSamples.kt
@@ -1,6 +1,10 @@
-package ch.tutteli.atrium.api.infix.en_GB.samples
+//TODO remove file with 1.0.0
+@file:Suppress("DEPRECATION")
+
+package ch.tutteli.atrium.api.infix.en_GB.samples.deprecated
 
 import ch.tutteli.atrium.api.infix.en_GB.*
+import ch.tutteli.atrium.api.infix.en_GB.samples.fails
 import ch.tutteli.atrium.api.verbs.internal.expect
 import ch.tutteli.atrium.translations.DescriptionComparableAssertion
 import kotlin.test.Test
